Modified main.rb
