+++ display_menu.rb
@@ -1,0 +1,12 @@
+class DisplayMenu
+def display_menu
+  puts 'Please choose an option by entering a number:'
+  puts '1. List all books'
+  puts '2. List all people'
+  puts '3. Create a person'
+  puts '4. Create a book'
+  puts '5. Create a rental'
+  puts '6. List rentals for a person'
+  puts '7. Exit'
+end
+end
